requirements.txt
requests
pyyaml
fake_useragent
tenacity
colorama
art
requests